+++ micro-model/bot/src/lib.rs
@@ -12,3 +12,4 @@ pub struct ComputeMove {
 }
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct MoveComputed(pub MakeMoveCommand);
+
