+++ tinybrain/src/katago/json.rs
@@ -1,8 +1,6 @@
 use crate::err::*;
-use micro_model_bot::MoveComputed;
 use micro_model_moves::*;
 use serde_derive::{Deserialize, Serialize};
-use std::convert::TryFrom;
 use std::str::FromStr;
 use uuid::Uuid;
 
@@ -87,22 +85,6 @@ impl AlphaNumOrPass {
         } else {
             Ok(AlphaNumOrPass(PASS.to_string()))
         }
-    }
-}
-
-impl TryFrom<KataGoResponse> for MoveComputed {
-    type Error = KataGoParseErr;
-    fn try_from(response: KataGoResponse) -> Result<Self, Self::Error> {
-        let game_id = response.game_id()?;
-        let player = response.player()?;
-        let coord: Option<Coord> = interpret_coord(&response.move_infos[0].r#move)?;
-        let req_id = ReqId(Uuid::new_v4());
-        Ok(MoveComputed(MakeMoveCommand {
-            game_id,
-            player,
-            coord,
-            req_id,
-        }))
     }
 }
 
@@ -328,25 +310,5 @@ mod tests {
     fn test_interpret_pass() {
         let actual = interpret_coord("pass");
         assert_eq!(actual.expect("parse"), None)
-    }
-
-    #[test]
-    fn move_computed_from() {
-        let actual = MoveComputed::try_from(KataGoResponse {
-            id: Id(format!("{}_1_WHITE", Uuid::nil().to_string())),
-            turn_number: 1,
-            move_infos: vec![MoveInfo {
-                r#move: "B3".to_string(),
-                order: 0,
-            }],
-        })
-        .expect("fail");
-        let expected = MoveComputed(MakeMoveCommand {
-            game_id: GameId(Uuid::nil()),
-            coord: Some(Coord { x: 1, y: 2 }),
-            player: Player::WHITE,
-            req_id: actual.0.req_id.clone(),
-        });
-        assert_eq!(actual, expected)
     }
 }
