+++ tinybrain/src/katago/mod.rs
@@ -70,6 +70,22 @@ pub fn start(move_computed_in: Sender<MoveComputed>, compute_move_out: Receiver<
     }
 }
 
+impl TryFrom<KataGoResponse> for MoveComputed {
+    type Error = crate::err::KataGoParseErr;
+    fn try_from(response: KataGoResponse) -> Result<Self, Self::Error> {
+        let game_id = response.game_id()?;
+        let player = response.player()?;
+        let coord: Option<Coord> = interpret_coord(&response.move_infos[0].r#move)?;
+        let req_id = ReqId(Uuid::new_v4());
+        Ok(MoveComputed(MakeMoveCommand {
+            game_id,
+            player,
+            coord,
+            req_id,
+        }))
+    }
+}
+
 fn launch_child() -> Result<Child, std::io::Error> {
     Command::new(PROGRAM)
         .arg(ARGS[0])
@@ -82,4 +98,29 @@ fn launch_child() -> Result<Child, std::io::Error> {
         .stdin(Stdio::piped())
         .stdout(Stdio::piped())
         .spawn()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use json::KataGoResponse;
+    #[test]
+    fn move_computed_from() {
+        let actual = MoveComputed::try_from(KataGoResponse {
+            id: Id(format!("{}_1_WHITE", Uuid::nil().to_string())),
+            turn_number: 1,
+            move_infos: vec![MoveInfo {
+                r#move: "B3".to_string(),
+                order: 0,
+            }],
+        })
+        .expect("fail");
+        let expected = MoveComputed(MakeMoveCommand {
+            game_id: GameId(Uuid::nil()),
+            coord: Some(Coord { x: 1, y: 2 }),
+            player: Player::WHITE,
+            req_id: actual.0.req_id.clone(),
+        });
+        assert_eq!(actual, expected)
+    }
 }
